FEAT / 하단바 정렬
