--- lib/pages/Community.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Community extends StatefulWidget {
-  @override
-  State<Community> createState() => _CommunityState();
-}
-
-class _CommunityState extends State<Community> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Text("Community"),
-    );
-  }
-}
