--- lib/pages/Map.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Map extends StatefulWidget {
-  @override
-  State<Map> createState() => _MapState();
-}
-
-class _MapState extends State<Map> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Text("Map"),
-    );
-  }
-}
