--- lib/pages/MyTrip.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MyTrip extends StatefulWidget {
-  @override
-  State<MyTrip> createState() => _MyTripState();
-}
-
-class _MyTripState extends State<MyTrip> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Text("MyTrip"),
-    );
-  }
-}
